backend/requirements.txt
anyio==3.7.0
asgiref==3.7.2
click==8.1.3
Django==4.2.2
django-cors-headers==4.2.0
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
fastapi==0.99.1
h11==0.14.0
httptools==0.5.0
idna==3.4
Pillow==10.0.0
pydantic==1.10.10
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.7.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
